+++ core/php/AbeilleParser.php
@@ -390,21 +390,20 @@
         $eq['modelForced'] = isset($eqModel['modelForced']) ? $eqModel['modelForced'] : false;
         if (isset($eqModel['modelPath'])) // Forced model variant case
             $eq['modelPath'] = $eqModel['modelPath'];
-
-        $private = $eqLogic->getConfiguration('ab::private', null);
-        if ($private !== null) {
-            $eq['private'] = $private;
+        if (isset($eqModel['private'])) {
+            $eq['private'] = $eqModel['private'];
             parserLog('debug', "  'private' updated to ".json_encode($eq['private'], JSON_UNESCAPED_SLASHES));
-        } else if (isset($eq['private']))
+        } else if ($GLOBALS['eqList'][$net][$addr]['private'])
             unset($GLOBALS['eqList'][$net][$addr]['private']);
+
         $fromDevice = $eqLogic->getConfiguration('ab::fromDevice', null); // OBSOLETE soon. Replaced by 'private'
         if ($fromDevice !== null) { // OBSOLETE soon. Replaced by 'private'
             $eq['fromDevice'] = $fromDevice;
             parserLog('debug', "  'fromDevice' updated to ".json_encode($eq['fromDevice']));
         } else if (isset($eq['fromDevice']))
             unset($GLOBALS['eqList'][$net][$addr]['fromDevice']);
-        $eq['tuyaEF00'] = $eqLogic->getConfiguration('ab::tuyaEF00', null); // OBSOLETE soon. Replaced by 'private'
-        parserLog('debug', "  'tuyaEF00' updated to ".json_encode($eq['tuyaEF00'])); // OBSOLETE soon. Replaced by 'private'
+        // $eq['tuyaEF00'] = $eqLogic->getConfiguration('ab::tuyaEF00', null); // OBSOLETE soon. Replaced by 'private'
+        // parserLog('debug', "  'tuyaEF00' updated to ".json_encode($eq['tuyaEF00'])); // OBSOLETE soon. Replaced by 'private'
         // $eq['xiaomi'] = $eqLogic->getConfiguration('ab::xiaomi', null); // OBSOLETE soon. Replaced by 'private'
         // parserLog('debug', "  'xiaomi' updated to ".json_encode($eq['xiaomi'])); // OBSOLETE soon. Replaced by 'private'
 
@@ -535,9 +534,9 @@
                 'modelSource' => '',
                 'modelForced' => isset($eqModel['modelForced']) ? $eqModel['modelForced'] : false,
                 'customization' => $eqLogic->getConfiguration('ab::customization', null),
-                //'private' => // Set if exists in model
+                //'private' => // Optional: Set if 'private' section exists in model
                 //'fromDevice' => // Set if exists in model // OBSOLETE soon
-                'tuyaEF00' => $eqLogic->getConfiguration('ab::tuyaEF00', null), // OBSOLETE soon => replaced by 'private'
+                // 'tuyaEF00' => $eqLogic->getConfiguration('ab::tuyaEF00', null), // OBSOLETE soon => replaced by 'private'
                 // 'xiaomi' => $eqLogic->getConfiguration('ab::xiaomi', null), // OBSOLETE soon => replaced by 'private'
                 // Optional 'notStandard-0400-0000'
             );