+++ docs/fr_FR/Changelog.rst
@@ -14,6 +14,7 @@ ChangeLog
 - Réseau: Affichage table des liens Zigate 1 à l'ouverture.
 - SonOff ZbminiL2: Correction 'mac capa' incorrect => 'RxOnWhenIdle'.
 - Interne: Mises-à-jour 'check_json' + 'update_json' pour changement 'tuyaEF00/fromDevice' => 'private/EF00'.
+- Modèles d'équipement: Suppression 'tuyaEF00/fromDevice' => remplacé par 'private/EF00'.
 
 240308-BETA-1
 -------------